pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudmark"
version = "0.1.0"
description = "A project for processing YAML files with TemplateURL."
authors = [{ name = "Tom Brookson", email = "tombrookson97@gmail.com" }]
dependencies = ["pyyaml>=5.4"]

[project.optional-dependencies]
dev = ["pytest>=6.2"]

[tool.pytest.ini_options]
testpaths = ["test"]
